Moved experimentutils to nupic.swarming
