+++ src/nupic/swarming/exp_generator/ExpGenerator.py
@@ -43,8 +43,7 @@ from nupic.frameworks.opf.metrics import MetricSpec
 from nupic.support import aggregationDivide
 from nupic.support.configuration import Configuration
 from nupic.support.enum import Enum
-from nupic.swarming.hypersearch.experimentutils import (InferenceType,
-                                           InferenceElement)
+from nupic.swarming.experimentutils import (InferenceType, InferenceElement)
