+++ src/nupic/swarming/exp_generator/descriptionTemplate.tpl
@@ -36,8 +36,7 @@ from nupic.frameworks.opf.expdescriptionhelpers import (
 
 from nupic.frameworks.opf.clamodelcallbacks import *
 from nupic.frameworks.opf.metrics import MetricSpec
-from nupic.swarming.hypersearch.experimentutils import (InferenceType,
-                                                        InferenceElement)
+from nupic.swarming.experimentutils import (InferenceType, InferenceElement)
 from nupic.support import aggregationDivide
 
 from nupic.frameworks.opf.opftaskdriver import (
