+++ src/nupic/swarming/experimentutils.py
